+++ www/static/js/dw-2.0.js
@@ -1488,7 +1488,7 @@ dw.chart = function(attributes) {
             var datasource;
 
             datasource = dw.datasource.delimited({
-                url: 'data',
+                url: 'data.csv',
                 firstRowIsHeader: chart.get('metadata.data.horizontal-header', true),
                 transpose: chart.get('metadata.data.transpose', false)
             });