+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 * added automatic number formatting
 * added column oauth_signin to user table (see migrate sql)
 * refactored JavaScript and CSS out of Twig templates
+* renamed chart data file to data.csv
 
 ### 1.5.4
 * added dw.utils.smartRound